+++ src/main/java/org/example/entities/Categoria.java
@@ -1,3 +1,5 @@
+package org.example.entities;
+
 import java.util.HashSet;
 import java.util.Set;
