+++ src/main/java/org/example/entities/Promocion.java
@@ -1,3 +1,5 @@
+package org.example.entities;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.HashSet;
