+++ src/main/java/org/example/entities/enums/FormaPago.java
@@ -1,3 +1,5 @@
+package org.example.entities.enums;
+
 public enum FormaPago {
     EFECTIVO,
     MERCADOPAGO
